Use input as filter for ssh and desktop connect menus (#47422)
Closes: https://github.com/gravitational/teleport/issues/41123

This will add an optional prop to denote a MenuLogin as "filter" instead
of a freeform input. If set to filter type, the input will only filter
the available logins and hitting "Enter" will now select the first item
in the filtered list. If there are no items, enter does nothing and no
longer allows any freeform entry.

Databases still allow freeform entry.

The other solution to this issue was to remove the input all together,
however I think have a filter input is better just in case a lot of
logins exist and customers are used to just typing in the login they
want (if it exists in the list of course)